% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{action}
\alias{action}
\title{Generates the outcome of an intervention or observation of a directed acyclic graph of binary variables}
\usage{
action(int, graph, br = 0.1, pow = 0.9)
}
\arguments{
\item{int}{is the intervention expressed as a vector with 1 as on, -1 as off and zero as free.
graph is directed acyclic graph expressed as a directed adjacency matrix
br is the background activation rate, defaults to .1
pow is the causal power of the links, defaults to .9}
}
\description{
This function computes the shannon entropy of a probability distribution.  It requires a second function called propagation.
}
\examples{
action(c(1,0,0), matrix(c(0,1,0,0,0,0,0,0,0),3,3, byrow=T))
}
\keyword{action}

