% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_int.nl.R
\name{choose_int.nl}
\alias{choose_int.nl}
\title{Get intervention values}
\usage{
choose_int.nl(DAGs, int_opts, prior, sar, pow, FUN, ...)
}
\arguments{
\item{DAGs}{an array where every slice is the next graph matrix in the hypothesis space}

\item{int_opts}{the available interventions in a maatrix where each row is an intervention 1 is on, -1 is off, 0 is free}

\item{prior}{the prior distribution over graphs}

\item{sar}{spontaneous activation rate}

\item{pow}{power of causal links}

\item{FUN}{the function to use to evaluate the intervention options relative to the hypothesis space and prior}
}
\description{
Computes the most useful intervention according to any desired function.
Does not require likelihoods to be preloaded as it uses indi_li to compute these on the fly
}
\examples{
int_opts<-as.matrix(expand.grid(rep(list(0:2), 3)))
int_opts[int_opts==2]<--1
out<-choose_int.nl(DAGs_3, int_opts, prior('random',25), .1, .8, shannon_entropy)
}
\keyword{expected}
\keyword{value}
