% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_bar.R
\name{error_bar}
\alias{error_bar}
\title{Creates error bars}
\usage{
error_bar(x, y, upper, lower = upper, length = 0.1, ...)
}
\arguments{
\item{x}{locations of bars (normally 1:n_bars).}

\item{y}{height of bars (pull this from the original plot by assigning the plot output to a variable).}

\item{upper}{extent of upper errorbar (e.g. compute standard errors and put these here).}

\item{lower}{extent of lower errorbar defaults to same as upper.}

\item{length}{width of bars}
}
\description{
This function adds error bars to a bar plot.
}
\examples{
barx<-barplot(1:3)
error_bar(barx, 1:3,c(.1,.1,.1))
}
\keyword{bars}
\keyword{error}
