% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{generate_unfolding}
\alias{generate_unfolding}
\title{Unfold a cyclic causal network}
\usage{
generate_unfolding(graph, starting_state, unfold_for = 20, pow = 0.75,
  sar = 0)
}
\arguments{
\item{graph}{The graph to be unfolded}

\item{starting_state}{The state at t=0}

\item{unfold_for}{How many steps to unfold the graph for? Defaults to 20}

\item{pow}{How likely are the causal connections to cause their effects, defaults to .75}

\item{sar}{How likely are the components to turn on on their own, defaults to 0.}
}
\description{
This function unfolds a cyclic causal graph probabilitically for a set number of time steps
}
\examples{
generate_unfolding(matrix(c(1,1,0,0), 2, 2, byrow=T), c(1,0))
}
\keyword{cyclic}
\keyword{networks}

