% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{get_joint}
\alias{get_joint}
\title{Get joint probability distribution over all possible unfoldings of a cyclic graph}
\usage{
get_joint(graph, starting_state, unfold_for = 2, pow = 0.75, sar = 0)
}
\arguments{
\item{graph}{the generating graph}

\item{starting_state}{the starting position or distribution for the chain}

\item{how}{many steps to unfold the joint NB Very slow for more than 5ish}

\item{pow}{strength of causal connections, defaults to .75}

\item{sar}{strength of background exogenous causes of variables, defaults to 0}
}
\description{
Returns the marginal probability of each state at time t and of each node being active at time t
}
\examples{
get_joint(matrix(c(1,1,0,0),2,2,byrow=T), c(1,0), unfold_for=7)
}
\keyword{distribution}
\keyword{joint}

