% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indi_like.R
\name{indi_like}
\alias{indi_like}
\title{Create likelihoods for a particular observation/intervention and graph on the fly}
\usage{
indi_like(graph, int, dat, sa, po, combination_function = "noisy_or")
}
\arguments{
\item{graph}{the true graph}

\item{int}{the intervention performed}

\item{dat}{the outcome observed}

\item{sa}{spontaneous activation rate}

\item{po}{power of causal links}

\item{combination_function}{the way in which multiple active causes interact, defaults to noisy_or, currently no other option}
}
\description{
An alternative to precomputing likelihood of all possible data patterns for when the problem is too large
}
\examples{
indi_like(DAGs_3[,,15], c(1,0,0), c(1,1,1), .1, .8)
}
\keyword{likelihoods}
