% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{initialise_payoffs}
\alias{initialise_payoffs}
\title{Create a value matrix for utility calculations}
\usage{
initialise_payoffs(r = 20, w = -10, m = -10, f = 0, b = 0)
}
\arguments{
\item{r}{Value link correctly endorsed, defaults to 20}

\item{w}{Value link incorrectly endorsed, defaults to -10}

\item{m}{Value link incorrectly not endorsed, defaults to -10}

\item{f}{Value link correctly not endorsed, defaults to 0}

\item{b}{Bonus completely correct, defaults to 0}
}
\description{
This function creates a value matrix for calculating the utility of endorsing a particular graph given your prior
Requires hypothesis graphs g to be in data.frame form in the environment
Requires valid proposal space vps to be in data.frame form in the environment too
}
\examples{
vps<-g
initialise_payoffs()
}
\keyword{utility}

