% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{likelihood}
\alias{likelihood}
\title{Create likelihood data.frame}
\usage{
likelihood(g, br = 0.1, pow = 0.9)
}
\arguments{
\item{g}{matrix of hypothesis graphs where each line is the graph adjacency marix written out by row}

\item{br}{power of background causes, defaults to .1}

\item{power}{of active endogenous causes, defaults to .8}
}
\description{
This creates a data-frame of likelihoods of outcomes under interventions on a DAG with a parameterisation
given noisy or function of active endogenous causes and constant exogenous causes
}
\examples{
li<-likelihood(g)
}
\keyword{likelihood}

