% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{likelihood_unfolding}
\alias{likelihood_unfolding}
\title{Get log likelihood of a particular unfolded graph}
\usage{
likelihood_unfolding(X, graph, pow = 0.75, sar = 0)
}
\arguments{
\item{X}{The unfolding}

\item{graph}{the graph for which to compute the likelihood of this unfolding}

\item{pow}{How likely are the causal connections to cause their effects, defaults to .75}

\item{sar}{How likely are the components to turn on on their own, defaults to 0.}
}
\description{
This function unfolds a cyclic causal graph probabilitically for a set number of time steps
}
\examples{
likelihood_unfolding(generate_unfolding(matrix(c(1,1,0,0), 2, 2, byrow=T), c(1,0)), matrix(c(1,1,0,0), 2, 2, byrow=T))
}
\keyword{cyclic}
\keyword{networks}

