% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{num2bin}
\alias{num2bin}
\title{Turn numbers to binary}
\usage{
num2bin(number, noBits)
}
\arguments{
\item{number}{the number}

\item{noBits}{the number of bits to include, if omitted defaults to 32}
}
\description{
This function just turns a number into a binary vector, useful for listing all permutations of two level conditions
}
\examples{
cat_function()
}
\keyword{bits}

