% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prog.R
\name{plotprog}
\alias{plotprog}
\title{A Cat Function}
\usage{
plotprog()
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}
