% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{posterior_pairwise}
\alias{posterior_pairwise}
\title{Posterior distribution over two graphs given prior and an unfolding}
\usage{
posterior_pairwise(graph1, graph2, unfolding, prior = c(0.5, 0.5),
  pow = 0.75, sar = 0)
}
\arguments{
\item{graph1}{The first hypothesis}

\item{graph2}{The second hypothesis}

\item{unfolding}{the data}

\item{prior}{the prior distribution (as a vector), defaults to c(.5,.5)}

\item{pow}{strength of causal links, defaults to .75}

\item{sar}{strengh of background causes, defaults to 0}
}
\description{
Posterior distribution over two graphs given prior and an unfolding
}
\examples{
out<-posterior_pairwise(matrix(c(1,1,0,0),2,2,byrow=T), matrix(c(1,0,0,0),2,2,byrow=T),
generate_unfolding(matrix(c(1,1,0,0),2,2,byrow=T), c(1,0)))
}
\keyword{likelihood}
\keyword{ratio}

