% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{prior}
\alias{prior}
\title{Generate a prior}
\usage{
prior(s, g)
}
\arguments{
\item{s}{what kind of prior do you want?  Can choose 'flat', 'random', 'flatSomeZero' and 'flatMostZero'}

\item{g}{matrix of hypothesis graphs, one written rowwise per row}
}
\description{
The function generates a prior in the form required for other functions, namely a column matrix
}
\examples{
prior('random', g)
}
\keyword{prior}

