% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propagation.R
\name{propagation}
\alias{propagation}
\title{Propagation function}
\usage{
propagation(gr, active, st, power)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function is a part of the process of calculating the outcome of a token intervention or observation
It makes little sense to include an example
}
\examples{
cat_function()
}
\keyword{cats}
