% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repmat.R
\name{repmat}
\alias{repmat}
\title{Replicates a matrix m by n times}
\usage{
repmat(X, m, n)
}
\arguments{
\item{X}{a matrix}

\item{m}{number of row replications}

\item{n}{number of row replications}
}
\description{
Replicates a matrix m by n times
}
\examples{
repmat(c(1,2,3), 2,5)

}
\keyword{matlab}
