% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resave.R
\name{resave}
\alias{resave}
\title{Resave a rData file with all the previous stuff plus more}
\usage{
resave(..., list = character(), file)
}
\arguments{
\item{file}{the file to add to}
}
\description{
Takes an exisiting rData file and adds to it
}
\examples{
resave(file='tmp_file.rdata', x=1)
}
\keyword{resave}
