% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannon_entropy.R
\name{shannon_entropy}
\alias{shannon_entropy}
\title{Computes Shannon entropy}
\usage{
shannon_entropy(pDist, log_type = "nats")
}
\arguments{
\item{pDist}{a probability distribution as a vector or a matrix of probability distributions columns.}

\item{log_type}{choose nats or bits, defaults to nats.}
}
\description{
This function computes the shannon entropy of a probability distribution.
}
\examples{
shannon_entropy(c(.5,.5), log_type='bits')
}
\keyword{entropy}
