% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{tsallis_entropy}
\alias{tsallis_entropy}
\title{Computes Tsallis entropy}
\usage{
tsallis_entropy(pDist, q)
}
\arguments{
\item{pDist}{the probability distribution in question}

\item{q}{the parameter that controls the function, support = [-Inf, Inf (q!=0)]}
}
\description{
This function computes the tsallis entropy of a distribution with a particular alpha function
Does not work for q=1
}
\examples{
tsallis_entropy(prior('random',g),2)
}
\keyword{entropy}

