% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_proposals.R
\name{valid_proposals}
\alias{valid_proposals}
\title{Valid proposal space for a causal judgement}
\usage{
valid_proposals()
}
\description{
This function creates the space of all reportable graphs i.e. all DAGs and all 3+ cycles for my experiments
Requires rowwise matrix g listing causal hypotheses g to be in the environment
}
\examples{
valid_proposals()
}
\keyword{causal}
\keyword{decision}
\keyword{judgment}
