% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{virility}
\alias{virility}
\title{Virility property of graphs}
\usage{
virility(graph)
}
\arguments{
\item{graph}{the graph to check}
}
\description{
How many forks and how many colliders does the graph have.
Returns number of forks - number of colliders
}
\examples{
virility(matrix(c(0,1,1,0,0,0,0,0,0),3,3, byrow=T))
}
\keyword{graph}

