% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{draw_graph}
\alias{draw_graph}
\title{Uses igraph to draw graphs how I like them}
\usage{
draw_graph(graph, labels)
}
\arguments{
\item{graph}{The adjacency matrix of the graph you want drawn}

\item{labels}{The names for the nodes in the graph}
}
\description{
Works well for 3-5 nodes
}
\examples{
draw_graph(matrix(c(0,1,1,0,0,0,0,0,0),3,3, byrow=T), c('A','B','C'))
}
\keyword{graph}

