% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{graph_cyclic}
\alias{graph_cyclic}
\title{Is this graph cyclic}
\usage{
graph_cyclic(graph)
}
\arguments{
\item{graph}{An adjacency matrix}
}
\description{
Does the graph contain a loop.  True if it does.
}
\examples{
graph_cyclic(matrix(c(0,1,1,0,0,0,0,0,0),3,3, byrow=T))
graph_cyclic(matrix(c(0,1,0,1,0,0,0,0,0),3,3, byrow=T))
}
\keyword{DAG}

