% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{my.graph.formula}
\alias{my.graph.formula}
\title{A slight mod of graph.formul() from igraphs to avoid the automatic removal of unconnected nodes}
\usage{
my.graph.formula(...)
}
\description{
A slight mod of graph.formul() from igraphs to avoid the automatic removal of unconnected nodes
See graph.formula for details
}
\examples{
g1<-my.graph.formula(A--+B, C)
plot(g1)
}
\keyword{igraph}

