% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{plotprog}
\alias{plotprog}
\title{A Cat Function}
\usage{
plotprog()
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}

