% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{prior}
\alias{prior}
\title{Generate a prior}
\usage{
prior(s, M)
}
\arguments{
\item{s}{what kind of prior do you want?  Can choose 'flat', 'random', 'flatSomeZero' and 'flatMostZero'}

\item{N}{number of hypotheses}
}
\description{
The function generates a prior in the form required for other functions, namely a column matrix
}
\examples{
prior('random', M)
}
\keyword{prior}
